+++ store/store.go
@@ -42,6 +42,23 @@ const (
 	appliedWaitDelay    = 100 * time.Millisecond
 )
 
+// QueryRequest represents a query that returns rows, and does not modify
+// the database.
+type QueryRequest struct {
+	Queries []string
+	Timings bool
+	Tx      bool
+	Lvl     ConsistencyLevel
+}
+
+// ExecuteRequest represents a query that returns now rows, but does modify
+// the database.
+type ExecuteRequest struct {
+	Queries []string
+	Timings bool
+	Tx      bool
+}
+
 // Transport is the interface the network service must provide.
 type Transport interface {
 	net.Listener
@@ -425,29 +442,32 @@ func (s *Store) Stats() (map[string]interface{}, error) {
 		return nil, err
 	}
 	status := map[string]interface{}{
-		"raft":          s.raft.Stats(),
-		"addr":          s.Addr().String(),
-		"leader":        s.Leader(),
-		"apply_timeout": s.ApplyTimeout.String(),
-		"meta":          s.meta,
-		"peers":         peers,
-		"dir":           s.raftDir,
-		"sqlite3":       dbStatus,
-		"db_conf":       s.dbConf,
+		"raft":               s.raft.Stats(),
+		"addr":               s.Addr().String(),
+		"leader":             s.Leader(),
+		"apply_timeout":      s.ApplyTimeout.String(),
+		"open_timeout":       s.OpenTimeout.String(),
+		"heartbeat_timeout":  s.HeartbeatTimeout.String(),
+		"snapshot_threshold": s.SnapshotThreshold,
+		"meta":               s.meta,
+		"peers":              peers,
+		"dir":                s.raftDir,
+		"sqlite3":            dbStatus,
+		"db_conf":            s.dbConf,
 	}
 	return status, nil
 }
 
 // Execute executes queries that return no rows, but do modify the database.
-func (s *Store) Execute(queries []string, timings, tx bool) ([]*sql.Result, error) {
+func (s *Store) Execute(ex *ExecuteRequest) ([]*sql.Result, error) {
 	if s.raft.State() != raft.Leader {
 		return nil, ErrNotLeader
 	}
 
 	d := &databaseSub{
-		Tx:      tx,
-		Queries: queries,
-		Timings: timings,
+		Tx:      ex.Tx,
+		Queries: ex.Queries,
+		Timings: ex.Timings,
 	}
 	c, err := newCommand(execute, d)
 	if err != nil {
@@ -499,16 +519,16 @@ func (s *Store) Backup(leader bool) ([]byte, error) {
 }
 
 // Query executes queries that return rows, and do not modify the database.
-func (s *Store) Query(queries []string, timings, tx bool, lvl ConsistencyLevel) ([]*sql.Rows, error) {
+func (s *Store) Query(qr *QueryRequest) ([]*sql.Rows, error) {
 	// Allow concurrent queries.
 	s.mu.RLock()
 	defer s.mu.RUnlock()
 
-	if lvl == Strong {
+	if qr.Lvl == Strong {
 		d := &databaseSub{
-			Tx:      tx,
-			Queries: queries,
-			Timings: timings,
+			Tx:      qr.Tx,
+			Queries: qr.Queries,
+			Timings: qr.Timings,
 		}
 		c, err := newCommand(query, d)
 		if err != nil {
@@ -531,11 +551,11 @@ func (s *Store) Query(queries []string, timings, tx bool, lvl ConsistencyLevel) 
 		return r.rows, r.error
 	}
 
-	if lvl == Weak && s.raft.State() != raft.Leader {
+	if qr.Lvl == Weak && s.raft.State() != raft.Leader {
 		return nil, ErrNotLeader
 	}
 
-	r, err := s.db.Query(queries, tx, timings)
+	r, err := s.db.Query(qr.Queries, qr.Tx, qr.Timings)
 	return r, err
 }
 
@@ -832,11 +852,7 @@ func (f *fsmSnapshot) Persist(sink raft.SnapshotSink) error {
 		}
 
 		// Close the sink.
-		if err := sink.Close(); err != nil {
-			return err
-		}
-
-		return nil
+		return sink.Close()
 	}()
 
 	if err != nil {