+++ system_test/helpers.go
@@ -73,6 +73,15 @@ func (n *Node) Query(stmt string) (string, error) {
 	return string(body), nil
 }
 
+// QueryMulti runs multiple queries against the node.
+func (n *Node) QueryMulti(stmts []string) (string, error) {
+	j, err := json.Marshal(stmts)
+	if err != nil {
+		return "", err
+	}
+	return n.postQuery(string(j))
+}
+
 // Join instructs this node to join the leader.
 func (n *Node) Join(leader *Node) error {
 	resp, err := DoJoinRequest(leader.APIAddr, n.RaftAddr)
@@ -84,6 +93,44 @@ func (n *Node) Join(leader *Node) error {
 	}
 	defer resp.Body.Close()
 	return nil
+}
+
+// Status returns the status and diagnostic output for node.
+func (n *Node) Status() (string, error) {
+	v, _ := url.Parse("http://" + n.APIAddr + "/status")
+
+	resp, err := http.Get(v.String())
+	if err != nil {
+		return "", err
+	}
+	if resp.StatusCode != 200 {
+		return "", fmt.Errorf("status endpoint returned: %s", resp.Status)
+	}
+	defer resp.Body.Close()
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", err
+	}
+	return string(body), nil
+}
+
+// Expvar returns the expvar output for node.
+func (n *Node) Expvar() (string, error) {
+	v, _ := url.Parse("http://" + n.APIAddr + "/debug/vars")
+
+	resp, err := http.Get(v.String())
+	if err != nil {
+		return "", err
+	}
+	if resp.StatusCode != 200 {
+		return "", fmt.Errorf("expvar endpoint returned: %s", resp.Status)
+	}
+	defer resp.Body.Close()
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", err
+	}
+	return string(body), nil
 }
 
 // ConfirmRedirect confirms that the node responds with a redirect to the given host.
@@ -108,6 +155,19 @@ func (n *Node) ConfirmRedirect(host string) bool {
 
 func (n *Node) postExecute(stmt string) (string, error) {
 	resp, err := http.Post("http://"+n.APIAddr+"/db/execute", "application/json", strings.NewReader(stmt))
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", err
+	}
+	return string(body), nil
+}
+
+func (n *Node) postQuery(stmt string) (string, error) {
+	resp, err := http.Post("http://"+n.APIAddr+"/db/query", "application/json", strings.NewReader(stmt))
 	if err != nil {
 		return "", err
 	}
@@ -254,6 +314,7 @@ func mustNewNode(enableSingle bool) *Node {
 	node.RaftAddr = node.Store.Addr().String()
 
 	node.Service = httpd.New("localhost:0", node.Store, nil)
+	node.Service.Expvar = true
 	if err := node.Service.Start(); err != nil {
 		node.Deprovision()
 		panic(fmt.Sprintf("failed to start HTTP server: %s", err.Error()))
@@ -301,4 +362,9 @@ func mustTempDir() string {
 		panic("failed to create temp dir")
 	}
 	return path
+}
+
+func isJSON(s string) bool {
+	var js map[string]interface{}
+	return json.Unmarshal([]byte(s), &js) == nil
 }